+++ java-io-base/src/main/java/internal/io/LockingFileFormatter.java
@@ -6,13 +6,13 @@ import nbbrd.io.WrappedIOException;
 import nbbrd.io.function.IOSupplier;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.channels.Channels;
 import java.nio.channels.FileChannel;
 import java.nio.channels.FileLock;
 import java.nio.channels.OverlappingFileLockException;
+import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 
@@ -25,12 +25,10 @@ public final class LockingFileFormatter<T> implements FileFormatter<T> {
 
     @Override
     public void formatFile(@NonNull T value, @NonNull File target) throws IOException {
-        try (FileOutputStream stream = new FileOutputStream(checkTarget(target))) {
-            try (FileLock ignore = stream.getChannel().lock()) {
-                delegate.formatStream(value, stream);
-            } catch (OverlappingFileLockException ex) {
-                throw WrappedIOException.wrap(ex);
-            }
+        try {
+            formatPath(value, target.toPath());
+        } catch (InvalidPathException ex) {
+            throw WrappedIOException.wrap(ex);
         }
     }