+++ java-io-base/src/main/java/internal/io/text/LegacyFiles.java
@@ -17,11 +17,13 @@
 package internal.io.text;
 
 import lombok.NonNull;
-import nbbrd.io.function.IOSupplier;
+import nbbrd.io.WrappedIOException;
+import org.checkerframework.checker.nullness.qual.Nullable;
 
 import java.io.*;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.nio.file.*;
 
 /**
  * @author Philippe Charles
@@ -29,54 +31,33 @@ import java.net.URISyntaxException;
 @lombok.experimental.UtilityClass
 public class LegacyFiles {
 
-    public static Reader openReader(CharSequence source) {
-        return new StringReader(source.toString());
+    public static @NonNull InputStream newInputStream(@NonNull File source) throws IOException {
+        return Files.newInputStream(FileSystemExceptions.checkSource(toPathOrRaiseIO(source)));
     }
 
-    public static Reader openReader(IOSupplier<? extends Reader> source) throws IOException {
-        return checkResource(source.getWithIO(), "Missing Reader");
+    public static @NonNull OutputStream newOutputStream(@NonNull File target) throws IOException {
+        return Files.newOutputStream(FileSystemExceptions.checkTarget(toPathOrRaiseIO(target)));
     }
 
-    public static InputStream openInputStream(IOSupplier<? extends InputStream> source) throws IOException {
-        return checkResource(source.getWithIO(), "Missing InputStream");
-    }
-
-    @NonNull
-    public static InputStream openInputStream(@NonNull File source) throws IOException {
-        return new BufferedInputStreamWithFile(FileSystemExceptions.checkSource(source));
-    }
-
-    public static Writer openWriter(IOSupplier<? extends Writer> source) throws IOException {
-        return checkResource(source.getWithIO(), "Missing Writer");
-    }
-
-    public static OutputStream openOutputStream(IOSupplier<? extends OutputStream> source) throws IOException {
-        return checkResource(source.getWithIO(), "Missing OutputStream");
-    }
-
-    @NonNull
-    public static OutputStream openOutputStream(@NonNull File target) throws IOException {
-        return new BufferedOutputStream(new FileOutputStream(FileSystemExceptions.checkTarget(target)));
+    public static @NonNull Path toPathOrRaiseIO(@NonNull File source) throws IOException {
+        try {
+            return source.toPath();
+        } catch (InvalidPathException ex) {
+            throw WrappedIOException.wrap(ex);
+        }
     }
 
     public String toSystemId(@NonNull File file) {
         return file.toURI().toASCIIString();
     }
 
-    public File fromSystemId(@NonNull String systemId) {
+    public @Nullable File fromSystemId(@NonNull String systemId) {
         if (systemId.startsWith("file:/")) {
             try {
-                return new File(new URI(systemId));
-            } catch (URISyntaxException ignore) {
+                return Paths.get(new URI(systemId)).toFile();
+            } catch (URISyntaxException | IllegalArgumentException | FileSystemNotFoundException ignore) {
             }
         }
         return null;
-    }
-
-    private static <T extends Closeable> T checkResource(T resource, String message) throws IOException {
-        if (resource == null) {
-            throw new IOException(message);
-        }
-        return resource;
     }
 }