+++ CHANGELOG.md
@@ -7,6 +7,18 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.0.33] - 2025-03-06
+
+### Added
+
+- Add static factory shortcuts to Picocsv parser/formatter
+- Add path shortcuts to CScriptWrapper and PowerShellWrapper
+- Add process shortcut to TextParser
+
+### Changed
+
+- Modernize use of NIO API
+
 ## [0.0.32] - 2024-09-25
 
 ### Fixed
@@ -280,7 +292,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-io-util/compare/v0.0.32...HEAD
+[Unreleased]: https://github.com/nbbrd/java-io-util/compare/v0.0.33...HEAD
+[0.0.33]: https://github.com/nbbrd/java-io-util/compare/v0.0.32...v0.0.33
 [0.0.32]: https://github.com/nbbrd/java-io-util/compare/v0.0.31...v0.0.32
 [0.0.31]: https://github.com/nbbrd/java-io-util/compare/v0.0.30...v0.0.31
 [0.0.30]: https://github.com/nbbrd/java-io-util/compare/v0.0.29...v0.0.30